Module Yocaml_git.Required

Interfaces required.

Requirements for building a Git Runtime.

module type SOURCE = sig ... end

Describes a natural transformation allowing a Yocaml program of type ‘a t to be transformed into a program of type ’a Lwt.t (so that it can be used via Yocaml_git.

module type CONFIG = sig ... end

Store configuration.